+++ backend/src/main/java/gr/ntua/ece/softeng35/backend/controllers/UsageTypeController.java
@@ -6,37 +6,60 @@ import org.springframework.web.bind.annotation.*;
 
 import gr.ntua.ece.softeng35.backend.models.UsageType;
 import gr.ntua.ece.softeng35.backend.models.UsageTypeRepository;
+import gr.ntua.ece.softeng35.backend.models.UserRepository;
 
 @RestController
 class UsageTypeController {
   private final UsageTypeRepository repository;
+  private final UserRepository repository2;
 
-  UsageTypeController(UsageTypeRepository repository) {
+  UsageTypeController(UsageTypeRepository repository, UserRepository repository2) {
     this.repository = repository;
+    this.repository2 = repository2;
   }
 
   @CrossOrigin(origins = "http://localhost:3000")
-  @GetMapping("/evcharge/api/admin/usagetypes")
-  List<UsageType> all() {
+  @GetMapping("/evcharge/api/admin/{apikey}/usagetypes")
+  List<UsageType> all(@PathVariable String apikey) {
+    CliController validator = new CliController(repository2);
+
+    if (!validator.validate(apikey)){
+      throw new NotAuthorizedException();
+    }
     return repository.findAll();
   }
 
   @CrossOrigin(origins = "http://localhost:3000")
-  @PostMapping("/evcharge/api/admin/usagetypesmod")
-  UsageType newUsageType(@RequestBody UsageType newUsageType) {
+  @PostMapping("/evcharge/api/admin/{apikey}/usagetypesmod")
+  UsageType newUsageType(@RequestBody UsageType newUsageType, @PathVariable String apikey) {
+    CliController validator = new CliController(repository2);
+
+    if (!validator.validate(apikey)){
+      throw new NotAuthorizedException();
+    }
     return repository.save(newUsageType);
   }
 
   @CrossOrigin(origins = "http://localhost:3000")
-  @GetMapping("/evcharge/api/admin/usagetypes/{id}")
-  UsageType one(@PathVariable Integer id) {
+  @GetMapping("/evcharge/api/admin/{apikey}/usagetypes/{id}")
+  UsageType one(@PathVariable Integer id, @PathVariable String apikey) {
+    CliController validator = new CliController(repository2);
+
+    if (!validator.validate(apikey)){
+      throw new NotAuthorizedException();
+    }
     return repository.findById(id)
       .orElseThrow(() -> new UsageTypeNotFoundException(id));
   }
 
   @CrossOrigin(origins = "http://localhost:3000")
-  @PutMapping("/evcharge/api/admin/usagetypesmod/{id}")
-  UsageType replaceUsageType(@RequestBody UsageType newUsageType, @PathVariable Integer id) {
+  @PutMapping("/evcharge/api/admin/{apikey}/usagetypesmod/{id}")
+  UsageType replaceUsageType(@RequestBody UsageType newUsageType, @PathVariable Integer id, @PathVariable String apikey) {
+    CliController validator = new CliController(repository2);
+
+    if (!validator.validate(apikey)){
+      throw new NotAuthorizedException();
+    }
     return repository.findById(id)
       .map(usageType -> {
         usageType.setTitle(newUsageType.getTitle());
@@ -47,8 +70,13 @@ class UsageTypeController {
   }
 
   @CrossOrigin(origins = "http://localhost:3000")
-  @DeleteMapping("/evcharge/api/admin/usagetypesmod/{id}")
-  void deleteUsageType(@PathVariable Integer id) {
+  @DeleteMapping("/evcharge/api/admin/{apikey}/usagetypesmod/{id}")
+  void deleteUsageType(@PathVariable Integer id, @PathVariable String apikey) {
+    CliController validator = new CliController(repository2);
+
+    if (!validator.validate(apikey)){
+      throw new NotAuthorizedException();
+    }
     repository.deleteById(id);
   }
 }
